GUI: (QIodeAbstractEditor) moved QIodeAbstractEditor class to tab_text_abstract.h/cpp files
