+++ gui/tabs/tab_report.cpp
@@ -1,5 +1,6 @@
 #include "tab_report.h"
 
+
 QIodeReportWidget::QIodeReportWidget(const QString& filepath, QTextEdit* output, std::shared_ptr<QIodeCompleter>& completer, 
     QWidget* parent) : QIodeAbstractEditor(I_REPORTS_FILE, filepath, parent)
 {
